core-primitives: add new crate sgx-runtime-primitives
