+++ app-libs/sgx-runtime/Cargo.toml
@@ -14,6 +14,9 @@ serde = { version = "1.0", optional = true, features = ["derive"] }
 codec = { package = "parity-scale-codec", version = "3.0.0", default-features = false, features = ["derive"] }
 scale-info = { version = "2.0.1", default-features = false, features = ["derive"] }
 
+# local dependencies
+itp-sgx-primitives = { path = "../../core-primitives/sgx-runtime-primitives", default-features = false }
+
 # Substrate dependencies
 frame-benchmarking = { optional = true, default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.28" }
 frame-executive = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.28" }
