+++ core-primitives/sgx-runtime-primitives/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "itp-sgx-primitives"
+version = "0.9.0"
+authors = ["Integritee AG <hello@integritee.network>"]
+edition = "2021"
+
+[dependencies]
+
+# Substrate dependencies
+sp-runtime = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.28" }
+sp-core = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.28" }
