+++ core-primitives/sgx-runtime-primitives/src/constants.rs
@@ -1,0 +1,12 @@
+use crate::types::{BlockNumber, Moment};
+
+pub const ONE_DAY: Moment = 86_400_000;
+
+pub const MILLISECS_PER_BLOCK: u64 = 6000;
+
+pub const SLOT_DURATION: u64 = MILLISECS_PER_BLOCK;
+
+// Time is measured by number of blocks.
+pub const MINUTES: BlockNumber = 60_000 / (MILLISECS_PER_BLOCK as BlockNumber);
+pub const HOURS: BlockNumber = MINUTES * 60;
+pub const DAYS: BlockNumber = HOURS * 24;
